pyproject.toml
[tool.poetry]
name = "bboom-task"
version = "0.1.0"
description = ""
authors = ["samsegomof <wemagonof91@gmail.com>"]
readme = "README.md"
packages = [{include = "bboom_task"}]

[tool.poetry.dependencies]
python = "^3.9"
poetry = "^1.5.1"
django = "4.0.5"
django-environ = "^0.10.0"
psycopg2-binary = "^2.9.6"
django-rest-framework = "^0.1.0"
drf-spectacular = "^0.26.4"
djangorestframework-simplejwt = "^5.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

